refactoring structure project
