+++ app/src/main/java/com/example/cocktailbar/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.cocktailbar
+package com.example.cocktailbar.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.cocktailbar.ui.theme.CocktailBarTheme
+import com.example.cocktailbar.presentation.ui.theme.CocktailBarTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
