+++ app/src/main/java/com/example/cocktailbar/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.cocktailbar.ui.theme
+package com.example.cocktailbar.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
